crawler/requirements.txt

# 크롤링
playwright==1.44.0
beautifulsoup4==4.12.3
instaloader==4.13

# HTTP
httpx==0.27.0

# ORM / DB
SQLAlchemy==2.0.30
PyMySQL==1.1.1
cryptography==42.0.8       # PyMySQL caching_sha2_password 지원

# 환경변수 / 스케줄링
python-dotenv==1.0.1
schedule==1.2.2

# 시간대 데이터 (Windows 필수)
tzdata>=2024.1
